pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "captacao-leads"
version = "0.1.0"
description = "Sistema de captação e disparo de leads B2B — Agência Simplifique"
requires-python = ">=3.11"
dependencies = [
    "apify-client>=1.8.0,<2.0",
    "firecrawl-py>=1.0.0,<2.0",
    "anthropic>=0.40.0,<1.0",
    "pydantic>=2.0,<3.0",
    "aio-pika>=9.0,<10.0",
    "python-dotenv>=1.0,<2.0",
    "structlog>=24.0,<25.0",
    "tenacity>=8.0,<9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["captacao"]

[tool.ruff.lint]
select = ["E", "F", "I"]
